+++ ts/components/screens/BaseScreenComponent.tsx
@@ -14,11 +14,12 @@ import * as React from "react";
 import { ColorValue, ModalBaseProps, Platform } from "react-native";
 import { TranslationKeys } from "../../../locales/locales";
 import {
-  instabugLog,
-  TypeLogs,
   openInstabugQuestionReport,
   openInstabugReplies,
-  DefaultReportAttachmentTypeConfiguration
+  DefaultReportAttachmentTypeConfiguration,
+  setInstabugSupportTokenAttribute,
+  TypeLogs,
+  instabugLog
 } from "../../boot/configureInstabug";
 import I18n from "../../i18n";
 import customVariables from "../../theme/variables";
@@ -34,6 +35,7 @@ import {
 import { SupportTokenState } from "../../store/reducers/authentication";
 import { getValueOrElse } from "../../features/bonus/bpd/model/RemoteValue";
 import { SupportRequestOptions } from "../SendSupportRequestOptions";
+import { SupportToken } from "../../../definitions/backend/SupportToken";
 import { AccessibilityEvents, BaseHeader } from "./BaseHeader";
 
 export interface ContextualHelpProps {
@@ -133,11 +135,25 @@ class BaseScreenComponent extends React.PureComponent<Props, State> {
     const maybeReport = this.state.requestReport;
     this.setState({ requestReport: none }, () => {
       maybeReport.map(type => {
-        fromNullable(this.state.supportToken)
-          .mapNullable(rsp => getValueOrElse(rsp, undefined))
-          .map(st => {
-            instabugLog(JSON.stringify(st), TypeLogs.INFO, "support-token");
-          });
+        const supportToken = fromNullable(this.state.supportToken)
+          .mapNullable<SupportToken | undefined>(rsp =>
+            getValueOrElse(rsp, undefined)
+          )
+          .getOrElse(undefined);
+
+        // Store/remove and log the support token only if is a new assistance request.
+        if (type === BugReporting.reportType.bug) {
+          // log on instabug the support token
+          if (supportToken) {
+            instabugLog(
+              JSON.stringify(supportToken),
+              TypeLogs.INFO,
+              "support-token"
+            );
+          }
+          // set or remove the support token as instabug user attribute
+          setInstabugSupportTokenAttribute(supportToken);
+        }
 
         switch (type) {
           case BugReporting.reportType.bug: